Added hook to protect /etc
With the  10-etc_guard.sh, top level files in /etc won't be override on
update.

This commit also fixed bugs where we don't re-compute the manifest file
after running the pre-install hook.

Also the manifest file is not sorted the opposite way (ascending) to
make it play nicely with comm