Conflicts: drastically speed up conflict detection
When checking for new files, didn't properly compare standardize the
output resulting in every file being treated as if it was new. This led
to expensive pkgmanager-owns calls which could take a while when there
is just a large file count.

Now we just use pkg_gen_manifest which was main to avoid this trap. Also
we check to ensure that the file in question is present and not a
directory and skip the call to the extension if this isn't true.