Make meta-packages work correctly with orphans
If the only package that had X as a dependencies is a meta-packages, X will
no longer be considered an orphan. The purpose of listing orphans is to
easily identify packages for removal. For example one may have installed
a few packages to test something and forgot the exact list. They could
run the orphans subcommand to see a list such packages. If all
intended/approved packages are kept is some arbitrary meta-packages, this
make picking the subcommand much more useful